+++ src/appAPI/index.ts
@@ -6,7 +6,7 @@ export default class AppAPI {
   credential: CredentialDTO;
   endpoints = {
     AUTH: '/auth',
-    PROJECT: '/projects',
+    PROJECT: '/project',
   };
 
   constructor(credential: CredentialDTO) {
@@ -41,6 +41,20 @@ export default class AppAPI {
         'Content-Type': 'application/json',
         'Authorization': `Bearer ${accessToken}`,
       },
+    });
+  }
+
+  private async postProject(accessToken: string, signal: AbortSignal, project: ProjectDTO) {
+    return fetch(this.credential.rootURL + this.endpoints.PROJECT, {
+      signal: signal,
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${accessToken}`,
+      },
+      body: JSON.stringify({
+        projects: [ project ],
+      }),
     });
   }
 
@@ -103,9 +117,69 @@ export default class AppAPI {
         return;
       }
 
-      // TODO: add the "status" propertie on received projects here as "new".
       onSuccess(body.projects);
-      // ===========================
+
+    } catch (error) {
+
+      if (error instanceof TypeError) {
+        switch (error.message) {
+          case 'Network request failed': onError('Network request failed. Did your phone or server lose internet connection?'); break;
+        }
+        return;
+      }
+
+      onError(
+        `${error}` +
+        `\n${JSON.stringify(error)}`
+      );
+    }
+  }
+
+  async uploadProject(
+    signal: AbortSignal,
+    projectDTO: ProjectDTO,
+    onSuccess: () => void,
+    onError: (error: string) => void,
+  ) {
+    try {
+
+      // AUTHENTICATION ===========================
+      const authResponse = await this.auth(signal);
+      if (!authResponse.ok) {
+        onError(
+          'The server did not recognize your credentials. Failed to authenticate.' +
+          '\nMethod: POST' +
+          '\nEndpoint: /auth' +
+          `\nStatus: ${authResponse.status}.`
+        );
+        return;
+      }
+
+      const authBody = await authResponse.json();
+      if (!authBody.accessToken) {
+        onError(
+          'Credentials accepted, but no AccessToken was found. Contact the developer of this server.' +
+          '\nMethod: POST' +
+          '\nEndpoint: /auth' +
+          `\nStatus: ${authResponse.status}` +
+          `\n{ ..., "accessToken: ${authBody.accessToken}" }`
+        );
+        return;
+      }
+
+      // POST PROJECT ======================================================
+      const response = await this.postProject(authBody.accessToken, signal, projectDTO);
+      if (!response.ok) {
+        onError(
+          'It was not possible to upload this project. The endpoint request failed. Contact the developer of this server.' +
+          '\nMethod: POST' +
+          '\nEndpoint: /project' +
+          `\nStatus: ${response.status}`
+        );
+        return;
+      }
+
+      onSuccess();
 
     } catch (error) {